extract product list mock to a separate file
This extracts the ProductSearchHit mock from the general products mock
file into "product-list.ts" file. This should allow us to add more mocks
 for the product list in the future and keeps the mocks for the product
 list separate from the product mocks.

I also extracted the utility function associated with this mock from
"utils.tsx" into "product-list.ts" as well. This is because the utility
 function is only used in the context of the product list mock.
